stats: extract PlannerStatsResolver from the orchestrator
Pure code motion, no behavior change. The planner-facing store
resolution — generation ordering, per-target completeness, the
pinned -> newest -> stale ladder, the planner stats cache with its
generation-scoped keyspaces and invalidation, and the lookup
diagnostics — moves into PlannerStatsResolver, one module with one
owner for the read policy. StatsOrchestrator keeps the capture policy
(sync-first bounded capture, async enqueue) and its public API, now
delegating store rungs to the resolver.

Motivation: the batch and single-target paths previously duplicated the
ladder shape, and a fallback fix landed on one path but not the other
until review caught it. One module holding both paths makes that class
of divergence structurally impossible.
